Change representation of rule tactics
A rule tactic (maybe better 'rule action'?) is now a MetaM tactic which
produces some additional information about what it did. For now, we
track dependent goals, i.e. goals produced by the rule tactic which
appear as metas in some of the other goals. This is in preparation for
proper metavariable tracking.