Store goal mvar dependencies, not occurrences
For any goal `g`, we used to store the set `g.mvars` of mvars which
occur in `g`. This is used to detect, for example, when `g` needs to be
copied because one of the metavariables in `g.mvars` is assigned.

However, for this purpose we need to store not only the mvars which
occur in `g` but also the mvars which occur in those mvars, etc. (i.e.
the mvars on which `g` depends).