DestructProducts: use transparency recursively
The `destructProducts` rule previously used the customisable
`destructProductsTransparency` only when first matching the type of a
hypothesis. Thus, for a hypothesis

  h : T ∧ U

with `T ≡ A ∧ B` semireducibly, `destructProducts` with
`destructProductsTransparency = reducible` would correctly match the
hypothesis (which is reducibly a product-like type), but would then
proceed to incorrectly match `T` and destruct it into `A` and `B`, even
though `T` is not reducibly a product-like type.

Now, `destructProducts` uses the `destructProductsTransparency` also
for recursive matching.