RuleExpr: remove checks for duplicate features
In attributes such as

    @[aesop safe unsafe ...]

the rule expression parser will now silently use the last phase
(`unsafe`) instead of throwing an error, and similar for features other
than the phase. This simplifies the code and may be useful for
macro-based abstration, where a macro gives default features that can
later be overridden.