RulePattern: fix defeq/syntactic eq mismatch
Patterns are elaborated with respect to the syntactic rule type, but
rule tactics use the rule type up to defeq (with customisable
transparency). We previously didn't take this mismatch fully into
account, so functions which work with the rule type (e.g. to instantiate
it with a pattern instantiation) could fail if the rule type had
'hidden' foralls.