Script: major refactor
This patch's major functional change is that rules may now produce
multiple script steps. These are 'atomic' tactics for which we store the
pre- and post-state. As a result, these tactics are structured
automatically by script structuring. Previously, we had to structure
them manually, and the results of this were imperfect.

We also redesign the core script types (particularly script builders)
accordingly and take this opportunity to split the Script file further.