BuiltinRules.Subst: hide forward implDetail hyps if they were revealed
The subst rule can turn `implDetail` hypotheses into regular ones via
the revert-intro pattern. As a result, it sometimes changes the kind of
`implDetail` hypotheses used by forward rules to 'default', which messes
with script generation. We add a hack to `subst` that makes the hyps
`implDetail`s again. This allows us to remove a hack when comparing
goals during script generation.