Support local forward rules (+ forward tracing)
Local forward rules are now re-elaborated each time we match a
hypothesis against them (and again when we reconstruct a queue entry).
Previously, they were elaborated once and for all for the first goal,
but this means that the resulting expressions could become invalid if
the internal names of hypotheses in the goal changed.