chore: modify `MonadStats` to not extend parents multiple times
MonadStats was extending `MonadLiftT` twice, but the two parents were defeq. This was a hack to get around the fact that the parent instances were providing `MonadLiftT` instances rather than a `MonadLift` instance, breaking `MonadLiftT` inference.

This changes it to *not* extend `MonadLiftT`, instead embedding it as a field, and then writing an explicit `MonadLift` instance.

This is motivated by [lean4#5770](https://github.com/leanprover/lean4/pull/5770), which will log warnings when structures extend parents multiple times.