Forward.State: add hyps lazily
We now delay adding hypotheses to the variable maps of cluster states
until at least one hypothesis is available for each slot. Doing so
prevents some unifications in case a slot is never matched. This seems
to be an important optimisation in practice for rules that have
unspecific both specific and non-specific slots.