Add incremental forward reasoning
We introduce a new, incremental forward reasoning algorithm. "Incremental"
means that the algorithm takes advantage of similarities between parent
and child goals, i.e. a forward rule that was run on the parent goal is
not re-run on the child goal unless the hypotheses that the rule was
applied to have changed or disappeared.

The new algorithm is enabled by default and can be toggled with the
`aesop.dev.statefulForward` option.

Breaking changes:

- Forward rules (including destruct rules) always use `reducible`
  transparency and no longer support custom indexing modes.
- `destruct` rules no longer clear non-propositional hypotheses.

See also
https://leanprover.zulipchat.com/#narrow/channel/270676-lean4/topic/Upcoming.20changes.20to.20Aesop.20forward.20rules

See the `forward-code` branch for the history of this feature.

closes #111