Fix forward rules with multiple complete matches
Previously, each forward match was translated into one rule. This meant
that when a forward rule had multiple complete matches for a goal, all
but one of these matches was ignored, since we implicitly treat the
applicable rules as a set.

To fix this, multiple matches for the same rule are now turned into one
rule that applies all the matches.

fixes #205