Ensure that auxiliary declarations on different branches have different names
We use the new `auxDeclNGen` infrastructure to ensure that each branch
of the search tree uses a separate name generator for auxiliary
declarations. This ensures that declaration names on different branches
don't conflict.

We delete `AesopTest.Environment`, which is superseded by the new (and
much more natural) `AesopTest.AuxDecl`, and
`AesopTest.NormSimpUnchangedMVarAssigned`, which got broken by recent
core changes.