Forward.State: remove hypTypes from ForwardState
This field used to be a hashset of RPINFs for all hypothesis types in
the ForwardState's corresponding goal. It was eagerly computed, which
is expensive in cases where we're using few forward rules, e.g. Mathlib.

The tradeoff for the removal of `hypTypes` is that when a forward rule
applies, it must linearly iterate through the entire local context to
determine whether the new hyp is redundant.