Index: stabilise ordering of index match locations and pattern substitutions
Prior to this commit, when a rule matched a goal with multiple index
match locations (e.g. multiple hypotheses) or multiple pattern
substitutions, the order of the locations/substitutions was partly based
on (hashes of) FVarId and MVarId names, hence unstable wrt the
'observable' local context.

To fix this, fvar match locations are now ordered by the fvar's index in
the current goal. Pattern substitutions are returned in lookup order by
the rule pattern discrimination tree, which is not ideal but better than
before.

fixes #264