RuleTac.Forward: use `synthAppInstances` for stateless forward reasoning
We now use `synthAppInstances` for both stateless and stateful forward
reasoning. Previously, stateless (but not stateful) used a simpler,
faster but less complete method that just tried to synth the instances
one by one, failing on the first failure. This is incomplete in the
presence of outparams.

I believe it should be possible to optimise `synthAppInstances` to
mitigate some of the performance impact, but for now the important part
is using the same method in both places.