Tree.Data.ForwardRuleMatches: fix erasure of consumed matches
The central function for updating forward rule matches had a bug that
led to consumed matches (those that we've already tried to apply,
successfully or unsuccessfully) not being removed. Hence, we would try
to apply the same matches potentially many times. This involves checking
whether a hypothesis of the resulting type already exists, which means
one defeq check for each hyp in the context.